+++ src/main/kotlin/app/revanced/patches/youtube/video/playerresponse/fingerprint/PlayerParameterBuilderFingerprint.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.video.videoid.fingerprint
+package app.revanced.patches.youtube.video.playerresponse.fingerprint
 
 import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
